Separate Core application part
Place models, content types, custom exceptions and history to separate project. Core is a UWP library because of UWP apps file access restrictions. They need Windows-specific tools (as StorageFile) for file access, and referenced .NET Standard libraries can access files via Stream passed from UWP layer only.
Models control data receiving. Sometimes we do not need to access the file system for that - for example, when data is cached - and then we do not need to create a stream. So it would be excessive to pass the stream to model from outside. That is why models need Windows.Storage and are placed in UWP library.